requirements.txt
docker-py
flask
configparser
